For separating Roll Numbers from names
The ERP has an option that lets you download the attendance sheet in a .csv format. However, the Roll numbers and names of the students all appear in a single cell, and cannot be separated easily. So, this Python script exactly does that. The output is a csv file containing the roll numbers and names separated, and only that. The attendance information is not preserved.
Input : export.csv
Output: Rolls_names.csv